clean up lockdown code
